# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for processing Excel order data",
        "title": "Excel Order API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/purchaseorders": {
            "get": {
                "description": "Retrieves a paginated list of purchase orders with optional search",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchaseorders"
                ],
                "summary": "Get list of purchase orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "pageNo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with list and metadata",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates new purchase orders and saves them to MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchaseorders"
                ],
                "summary": "Create purchase orders",
                "parameters": [
                    {
                        "description": "List of purchase orders to create",
                        "name": "orders",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PurchaseOrder"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/purchaseorders/import": {
            "post": {
                "description": "Retrieves purchase order data from an uploaded Excel file and returns it in JSON format",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchaseorders"
                ],
                "summary": "Import purchase orders from Excel file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel file containing purchase order data",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PurchaseOrder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - missing or invalid file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.PurchaseOrder": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "customer_po": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "integer"
                },
                "discount_percent": {
                    "type": "integer"
                },
                "extend_list_price": {
                    "type": "integer"
                },
                "extend_net_price": {
                    "type": "integer"
                },
                "extend_unit_net_price": {
                    "type": "integer"
                },
                "job_amount": {
                    "type": "integer"
                },
                "job_id_no": {
                    "type": "string"
                },
                "ordered": {
                    "type": "integer"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "product_code": {
                    "type": "string"
                },
                "product_description": {
                    "type": "string"
                },
                "project_manager": {
                    "type": "string"
                },
                "purchasing": {
                    "type": "string"
                },
                "received": {
                    "type": "integer"
                },
                "remain": {
                    "type": "integer"
                },
                "sales_team": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit_list_price": {
                    "type": "integer"
                }
            }
        }
    }
}